% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greg_all.R
\name{greg_all}
\alias{greg_all}
\title{greg_all}
\usage{
greg_all(plot_df, estimation, pixel_estimation_means, formula)
}
\arguments{
\item{plot_df}{A data frame containing the response variable, predictors, estimation unit, and resolution unit for each "plot"}

\item{estimation}{A character specifying the estimation column name within the other dataframes}

\item{pixel_estimation_means}{A dataframe with a column for the estimation unit and a column for the mean response variable value per that estimation unit}

\item{formula}{Formula to be used for the model, names should be consistent with the column names in plot_df and pixel_estimation_means}
}
\description{
This function runs the Generalized Regression Operating on Resolutions of Y estimator, also know as GREGORY, on a set of data.
}
\keyword{forest}
