% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregory_all.R
\name{gregory_all}
\alias{gregory_all}
\title{gregory_all}
\usage{
gregory_all(
  plot_df,
  resolution,
  estimation,
  pixel_estimation_means,
  proportions,
  formula,
  prop
)
}
\arguments{
\item{plot_df}{A data frame containing the response variable, predictors, estimation unit, and resolution unit for each "plot"}

\item{resolution}{A character specifying the resolution column name within the other dataframes}

\item{estimation}{A character specifying the estimation column name within the other dataframes}

\item{pixel_estimation_means}{A dataframe with a column for the estimation unit and a column for the mean response variable value per that estimation unit}

\item{proportions}{A dataframe with three columns: one for resolution, one for estimation, and one for the proportion of a resolution area found in each estimation area}

\item{formula}{Formula to be used for the model, names should be consistent with the column names in plot_df and pixel_estimation_means}

\item{prop}{A character specifying the column name of the proportion found in proportions}
}
\description{
This function runs the Generalized Regression Operating on Resolutions of Y estimator, also know as GREGORY, on a set of data.
}
\keyword{forest}
