% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregory_all.R
\name{gregory_all}
\alias{gregory_all}
\title{gregory_all}
\usage{
gregory_all(
  plot_df,
  resolution,
  estimation,
  pixel_estimation_means,
  proportions,
  formula,
  prop
)
}
\arguments{
\item{plot_df}{blah}

\item{resolution}{nlahh}

\item{estimation}{Nhahuahuauh}

\item{pixel_estimation_means}{uuurrrfgfghh}

\item{proportions}{oooouuuuhhhh}

\item{formula}{auauahhuuyyuuu}

\item{prop}{huuuyyyyyyyeeeeeuuurrrgh}
}
\description{
This runs gregory yay
}
\keyword{forest}
